requirements.txt
amqp==5.1.1
asgiref==3.5.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
bizerror==0.4.9
celery==5.2.7
certifi==2020.6.20
chardet==3.0.4
click==8.1.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
decorator==5.1.1
dictop==0.2.1
Django==3.2
django-behaviors==0.5.1
django-changelist-toolbar-admin==0.4.0
django-ckeditor==6.0.0
django-cors-headers==3.11.0
django-file-resubmit==0.5.2
django-filter==2.4.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-modelcluster==5.1
django-multiselectfield==0.1.12
django-mysql==3.11.1
django-phonenumber-field==7.0.0
django-selectel-storage==1.0.2
django-simple-export-admin==0.1.6
django-static-fontawesome==5.15.4.0
django-taggit==1.3.0
djangorestframework==3.12.1
draftjs-exporter==2.1.7
environs==9.2.0
et-xmlfile==1.1.0
fastutils==0.42.14
gunicorn==20.0.4
html5lib==1.1
idna==2.10
ImageHash==4.3.1
jdcal==1.4.1
jsonformatter==0.3.1
kombu==5.2.4
lxml==4.6.3
magic-import==0.2.0
marshmallow==3.9.1
msgpack==1.0.4
mysqlclient==2.0.3
numpy==1.20.1
openpyxl==3.0.10
phonenumbers==8.13.0
Pillow==8.0.1
prompt-toolkit==3.0.8
psutil==5.9.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycryptodome==3.15.0
Pygments==2.8.1
pypinyin==0.47.1
pypng==0.20220715.0
python-barcode==0.13.1
python-docx==0.8.10
python-dotenv==0.15.0
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
qrcode==7.4.2
redis==3.5.3
reportlab==3.5.55
requests==2.24.0
rsa==4.9
scipy==1.10.1
six==1.15.0
sorl-thumbnail==12.8.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
tokenize-rt==5.0.0
typing-extensions==3.7.4.3
Unidecode==1.1.1
urllib3==1.25.11
validators==0.18.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.4
xlrd==1.2.0
xlsxhelper==0.3.1
XlsxWriter==1.3.7
xlwt==1.3.0
